Correct Bottom Navigation bar implementation and add items to nav
