+++ lib/pages/feed.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class Feed extends StatefulWidget {
+  @override
+  _FeedState createState() => _FeedState();
+}
+
+class _FeedState extends State<Feed> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: Text('Feed Page'),
+      ),
+    );
+  }
+}
