+++ lib/pages/issues.dart
@@ -3,7 +3,6 @@ import '../pages/issue_detail.dart';
 
 import '../services/api.dart';
 import '../data/models.dart';
-import 'login_signup.dart';
 
 class PaginatedClass extends StatefulWidget {
   @override
@@ -19,15 +18,8 @@ class PaginatedClassState extends State<PaginatedClass>
   late Future _getObj;
   late String paginatedUrl;
 
-  int _selectedIndex = 0;
   static const TextStyle optionStyle =
       TextStyle(fontSize: 30, fontWeight: FontWeight.bold);
-
-  void _onItemTapped(int index) {
-    setState(() {
-      _selectedIndex = index;
-    });
-  }
 
   @override
   void initState() {
@@ -84,74 +76,6 @@ class PaginatedClassState extends State<PaginatedClass>
   Widget build(BuildContext context) {
     return SafeArea(
       child: Scaffold(
-        appBar: AppBar(
-          centerTitle: true,
-          title: Image.asset(
-            'assets/bugheist_logo.png',
-            fit: BoxFit.cover,
-            height: 30,
-          ),
-          actions: <Widget>[
-            IconButton(
-              icon: Icon(
-                Icons.search,
-                color: Colors.black,
-              ),
-              onPressed: () {
-                // do something
-              },
-            ),
-            IconButton(
-              icon: Icon(
-                Icons.portrait_rounded,
-                color: Colors.black,
-              ),
-              onPressed: () {
-                // do something
-              },
-            )
-          ],
-        ),
-        drawer: Drawer(
-          // Add a ListView to the drawer. This ensures the user can scroll
-          // through the options in the drawer if there isn't enough vertical
-          // space to fit everything.
-          child: ListView(
-            // Important: Remove any padding from the ListView.
-            padding: EdgeInsets.zero,
-            children: <Widget>[
-              DrawerHeader(
-                child: Text('Drawer Header'),
-                decoration: BoxDecoration(
-                  color: Color(0xFFDC4654),
-                ),
-              ),
-              ListTile(
-                title: Text('Login'),
-                onTap: () {
-                  // Update the state of the app
-                  // ...
-                  // Then close the drawer
-                  Navigator.pushReplacement(
-                    context,
-                    MaterialPageRoute(
-                      builder: (context) => LoginSignUp(),
-                    ),
-                  );
-                },
-              ),
-              ListTile(
-                title: Text('Item 2'),
-                onTap: () {
-                  // Update the state of the app
-                  // ...
-                  // Then close the drawer
-                  Navigator.pop(context);
-                },
-              ),
-            ],
-          ),
-        ),
         body: FutureBuilder(
           future: _getObj,
           builder: (context, snapshot) {
@@ -282,30 +206,6 @@ class PaginatedClassState extends State<PaginatedClass>
               ),
             );
           },
-        ),
-        bottomNavigationBar: BottomNavigationBar(
-          type: BottomNavigationBarType.fixed,
-          items: const <BottomNavigationBarItem>[
-            BottomNavigationBarItem(
-              icon: Icon(Icons.home),
-              label: 'Home',
-            ),
-            BottomNavigationBarItem(
-              icon: Icon(Icons.bug_report),
-              label: 'Report',
-            ),
-            BottomNavigationBarItem(
-              icon: Icon(Icons.list),
-              label: 'Issues',
-            ),
-            BottomNavigationBarItem(
-              icon: Icon(Icons.leaderboard),
-              label: 'Leaderboard',
-            ),
-          ],
-          currentIndex: _selectedIndex,
-          selectedItemColor: Color(0xFFDC4654),
-          onTap: _onItemTapped,
         ),
       ),
     );
