+++ lib/pages/leaderboard.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class LeaderBoard extends StatefulWidget {
+  @override
+  _LeaderBoardState createState() => _LeaderBoardState();
+}
+
+class _LeaderBoardState extends State<LeaderBoard> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(child: Text(
+        'Leaderboard Page'
+      ),),
+    );
+  }
+}
