+++ lib/pages/report_bug.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class ReportBug extends StatefulWidget {
+  @override
+  _ReportBugState createState() => _ReportBugState();
+}
+
+class _ReportBugState extends State<ReportBug> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: Text('Report a Bug'),
+      ),
+    );
+  }
+}
